requirements.txt
streamlit
pandas
sentence-transformers
faiss-cpu
numpy
scipy
tqdm
requests
torch
transformers
tokenizers
